requirements.txt

cryptography
PyGitHub
python-gitlab
atlassian-python-api
python-dotenv
ConfigArgParse
pytest
gitpython